move the panel and header to feature modal
* This is done to avoid the sparks notification to trigger again when going back to the suggestion from the outline.
* Also includes a fix to the usage counter tooltip
* Fixes missing focus on the focus keyphrase in ai generator
