+++ packages/js/src/ai-content-planner/components/content-suggestions-modal.js
@@ -1,17 +1,13 @@
-import { Badge, Link, Modal, SkeletonLoader, useSvgAria, Notifications } from "@yoast/ui-library";
+import { Modal, SkeletonLoader, useSvgAria } from "@yoast/ui-library";
 import { __ } from "@wordpress/i18n";
-import { ReactComponent as YoastIcon } from "../../../images/Yoast_icon_kader.svg";
 import { ReactComponent as Yoast } from "../../../images/yoast.svg";
-import { UsageCounter } from "@yoast/ai-frontend";
-import { QuestionMarkCircleIcon } from "@heroicons/react/solid";
 import { noop } from "lodash";
-import { useRef, useEffect, useCallback } from "@wordpress/element";
+import { useCallback } from "@wordpress/element";
 import { Transition } from "@headlessui/react";
 import { ContentPlannerError } from "./content-planner-error";
 import { IntentBadge } from "./intent-badge";
 import { ASYNC_ACTION_STATUS } from "../../shared-admin/constants";
 import { useFetchContentSuggestions } from "../hooks";
-import { SparksLimitNotification } from "../../ai-generator/components/sparks-limit-notification";
 
 /**
  * @typedef {import( "../constants" ).Suggestion} Suggestion
@@ -207,91 +203,31 @@ const ModalBodyContent = ( { status, suggestions, onSuggestionClick, error, onRe
  *
  * @param {Object} props The component props.
  * @param {string} props.status The current modal status.
- * @param {boolean} props.isPremium Whether the user has a premium add-on activated or not.
  * @param {Function} props.onSuggestionClick The function to call when a suggestion is clicked.
  * @param {Suggestion[]} props.suggestions The list of content suggestions to display.
  * @param {boolean} props.skipTransitions Whether to skip transition animations.
- * @param {number} props.usageCount The number of times the user has used the content suggestions feature.
- * @param {number} props.usageCountLimit The maximum number of times the user can use the content suggestions feature before hitting a limit.
- * @param {string} props.usageCountStatus The status of the usage count (e.g. "approaching", "reached").
  *
  * @returns {JSX.Element} The ContentSuggestionsModal component.
  */
 export const ContentSuggestionsModal = ( {
 	status,
-	isPremium,
 	onSuggestionClick = noop,
 	suggestions,
 	skipTransitions = false,
-	usageCount,
-	usageCountLimit,
-	usageCountStatus,
 	error,
-	modalHelpLink,
 } ) => {
-	const svgAriaProps = useSvgAria();
-	const closeButtonRef = useRef( null );
-
-	useEffect( () => {
-		closeButtonRef.current?.focus();
-	}, [ status ] );
-
 	const fetchContentSuggestions = useFetchContentSuggestions();
 
 	return (
-		<Modal.Panel
-			className="yst-p-0 yst-max-w-2xl yst-overflow-visible"
-			hasCloseButton={ false }
-		>
-			<Modal.CloseButton ref={ closeButtonRef } screenReaderText={ __( "Close content suggestions modal", "wordpress-seo" ) } />
-			<Modal.Container>
-				<Modal.Container.Header className="yst-flex yst-items-center yst-gap-2 yst-pe-12 yst-py-6 yst-ps-6 yst-border-b yst-border-slate-200">
-					<YoastIcon className="yst-fill-primary-500 yst-w-4 yst-mb-[1px]" { ...svgAriaProps } />
-					<Modal.Title size="2">{ __( "Content suggestions", "wordpress-seo" ) }</Modal.Title>
-					<Link
-						href={ modalHelpLink }
-						variant="primary"
-						className="yst-no-underline"
-						target="_blank"
-						rel="noopener noreferrer"
-						aria-label={ __( "Learn more about AI (Opens in a new browser tab)", "wordpress-seo" ) }
-					>
-						<QuestionMarkCircleIcon { ...svgAriaProps } className="yst-w-4 yst-h-4 yst-text-slate-500 yst-shrink-0" />
-					</Link>
-					<span className="yst-flex-grow" />
-					<Badge size="small">{ __( "Beta", "wordpress-seo" ) }</Badge>
-					{ status !== ASYNC_ACTION_STATUS.error && (
-						<UsageCounter
-							limit={ usageCountLimit }
-							requests={ usageCount }
-							mentionBetaInTooltip={ isPremium }
-							mentionResetInTooltip={ isPremium }
-							isSkeleton={ status === ASYNC_ACTION_STATUS.loading || usageCountStatus === ASYNC_ACTION_STATUS.loading }
-						/>
-					) }
-				</Modal.Container.Header>
-				<Modal.Container.Content className="yst-overflow-y-auto yst-p-6 yst-m-0">
-					<ModalBodyContent
-						status={ status }
-						suggestions={ suggestions }
-						onSuggestionClick={ onSuggestionClick }
-						error={ error }
-						onRetry={ fetchContentSuggestions }
-						skipTransitions={ skipTransitions }
-					/>
-				</Modal.Container.Content>
-			</Modal.Container>
-			<Notifications
-				className={
-					// Margin tricks to break out of the container. Transition to prevent sudden location jumps when loading new suggestions.
-					"yst-mx-[calc(50%-50vw)] yst-transition-all"
-				}
-				position="bottom-left"
-			>
-				{ status !== ASYNC_ACTION_STATUS.loading && (
-					<SparksLimitNotification className="yst-mx-[calc(50%-50vw)] yst-transition-all" />
-				) }
-			</Notifications>
-		</Modal.Panel>
+		<Modal.Container.Content className="yst-overflow-y-auto yst-p-6 yst-m-0">
+			<ModalBodyContent
+				status={ status }
+				suggestions={ suggestions }
+				onSuggestionClick={ onSuggestionClick }
+				error={ error }
+				onRetry={ fetchContentSuggestions }
+				skipTransitions={ skipTransitions }
+			/>
+		</Modal.Container.Content>
 	);
 };
