+++ packages/js/src/ai-content-planner/containers/content-outline-modal.js
@@ -2,7 +2,6 @@ import { compose } from "@wordpress/compose";
 import { withSelect, withDispatch } from "@wordpress/data";
 import { ContentOutlineModal } from "../components/content-outline-modal";
 import { CONTENT_PLANNER_STORE, FEATURE_MODAL_STATUS } from "../constants";
-import { STORE_NAME_AI } from "../../ai-generator/constants";
 
 export default compose( [
 	withSelect( ( select ) => {
@@ -14,23 +13,15 @@ export default compose( [
 			selectContentOutlineError,
 		} = select( CONTENT_PLANNER_STORE );
 
-		const {
-			selectUsageCountLimit,
-			selectUsageCount,
-		} = select( STORE_NAME_AI );
-
-		const { getIsPremium, selectLink } = select( "yoast-seo/editor" );
+		const { getIsPremium } = select( "yoast-seo/editor" );
 
 		return {
 			suggestion: selectSuggestion(),
 			outline: selectContentOutline(),
-			sparksLimit: selectUsageCountLimit(),
-			sparksUsage: selectUsageCount(),
 			status: selectContentOutlineStatus(),
 			error: selectContentOutlineError(),
 			isPremium: getIsPremium(),
 			isActive: selectFeatureModalStatus() === FEATURE_MODAL_STATUS.contentOutline,
-			modalHelpLink: selectLink( "https://yoa.st/ai-content-planner-help-button-modal" ),
 		};
 	} ),
 	withDispatch( ( dispatch ) => {
