+++ packages/js/tests/ai-content-planner/components/content-outline-modal.test.js
@@ -3,11 +3,6 @@ import { Modal } from "@yoast/ui-library";
 import { ContentOutlineModal } from "../../../src/ai-content-planner/components/content-outline-modal";
 import { ASYNC_ACTION_STATUS } from "../../../src/shared-admin/constants";
 
-const mockUsageCounter = jest.fn( () => null );
-jest.mock( "@yoast/ai-frontend", () => ( {
-	UsageCounter: ( props ) => mockUsageCounter( props ),
-} ) );
-
 jest.mock( "@wordpress/data", () => ( {
 	useSelect: jest.fn( () => false ),
 } ) );
@@ -59,8 +54,6 @@ const renderModal = ( { onClose = jest.fn(), status = ASYNC_ACTION_STATUS.loadin
 				suggestion={ defaultSuggestion }
 				status={ status }
 				isPremium={ false }
-				sparksLimit={ undefined }
-				sparksUsage={ undefined }
 				isActive={ false }
 				{ ...props }
 			/>
@@ -84,7 +77,6 @@ const renderLoadedModal = ( props ) => {
 
 describe( "ContentOutlineModal", () => {
 	beforeEach( () => {
-		mockUsageCounter.mockClear();
 		jest.useFakeTimers();
 	} );
 
@@ -92,50 +84,7 @@ describe( "ContentOutlineModal", () => {
 		jest.useRealTimers();
 	} );
 
-	describe( "close button", () => {
-		it( "calls onClose when the close button is clicked", () => {
-			const onClose = jest.fn();
-			renderModal( { onClose } );
-			fireEvent.click( screen.getByRole( "button", { name: /close/i } ) );
-			expect( onClose ).toHaveBeenCalledTimes( 1 );
-		} );
-	} );
-
-	describe( "header", () => {
-		it( "shows the 'Content outline' title", () => {
-			renderModal();
-			expect( screen.getByText( "Content outline" ) ).toBeInTheDocument();
-		} );
-
-		it( "shows the 'Beta' badge", () => {
-			renderModal();
-			expect( screen.getByText( "Beta" ) ).toBeInTheDocument();
-		} );
-
-		it( "shows the UsageCounter when sparksLimit is provided", () => {
-			renderModal( { sparksLimit: 10, sparksUsage: 3 } );
-			expect( mockUsageCounter ).toHaveBeenCalledWith(
-				expect.objectContaining( { limit: 10, requests: 3 } )
-			);
-		} );
-	} );
-
 	describe( "accessibility", () => {
-		it( "has a descriptive close button label", () => {
-			renderModal();
-			expect( screen.getByRole( "button", { name: "Close content outline" } ) ).toBeInTheDocument();
-		} );
-
-		it( "has an accessible dialog name from the title", () => {
-			renderModal();
-			expect( screen.getByRole( "dialog", { name: "Content outline" } ) ).toBeInTheDocument();
-		} );
-
-		it( "has an accessible description for the modal", () => {
-			renderModal();
-			expect( screen.getByText( "Review and customize your content outline before adding it to your post" ) ).toBeInTheDocument();
-		} );
-
 		it( "does not show the structure section when loading", () => {
 			renderModal();
 			act( () => {
@@ -413,11 +362,6 @@ describe( "ContentOutlineModal", () => {
 		it( "does not render the Add outline to post footer button while in error", () => {
 			renderErrorModal();
 			expect( screen.queryByRole( "button", { name: "Add outline to post" } ) ).not.toBeInTheDocument();
-		} );
-
-		it( "does not render the UsageCounter while in error", () => {
-			renderErrorModal( { sparksLimit: 10, sparksUsage: 2 } );
-			expect( mockUsageCounter ).not.toHaveBeenCalled();
 		} );
 
 		it( "calls onRetry when the retry button is clicked", () => {
